fix layerwise benchmark
